+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 ## v3.0.0
 
-- Use Node.js 12 as it is supported on Windows7
+- Use Node12 as it is supported on Windows7
 
 
 ## v2.0.0